GPL page code  refractor
