+++ SparkyFitnessFrontend/src/pages/Medications/Glp1Coach.tsx
@@ -1,4 +1,4 @@
-import { useMemo, useState } from 'react';
+import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import {
   Area,
@@ -10,50 +10,22 @@ import {
   XAxis,
   YAxis,
 } from 'recharts';
-import { Syringe, AlertTriangle, Plus, Trash2 } from 'lucide-react';
-import { INJECTION_SITES, todayInZone, addDays } from '@workspace/shared';
-import InjectionSiteBodyMap from './InjectionSiteBodyMap';
-import InjectionSiteSettings from './InjectionSiteSettings';
-import FastingTimer from './FastingTimer';
-import { Button } from '@/components/ui/button';
+import { useSerumCurve } from '@/hooks/useMedications';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Badge } from '@/components/ui/badge';
-import { Label } from '@/components/ui/label';
-import { Input } from '@/components/ui/input';
-import { Switch } from '@/components/ui/switch';
-import { Textarea } from '@/components/ui/textarea';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
-import {
-  Dialog,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/dialog';
-import {
-  useSerumCurve,
-  useSiteSuggestion,
-  useMedicationPens,
-  useMedicationInjections,
-  useMedicationTitration,
-  useLogInjectionMutation,
-  useCreatePenMutation,
-  useAddTitrationStepMutation,
-  useDeleteTitrationStepMutation,
-  useDeleteInjectionMutation,
-} from '@/hooks/useMedications';
-import { usePreferences } from '@/contexts/PreferencesContext';
-import type { Medication, MedicationPen } from '@/types/medications';
+import type { Medication } from '@/types/medications';
+import FastingTimer from './FastingTimer';
+import Glp1LogInjection from './Glp1LogInjection';
+import Glp1InventoryManager from './Glp1InventoryManager';
+import Glp1TitrationManager from './Glp1TitrationManager';
+import Glp1RecentInjections from './Glp1RecentInjections';
 
-export default function Glp1Coach({ med }: { med: Medication }) {
+interface Glp1CoachProps {
+  med: Medication;
+}
+
+export default function Glp1Coach({ med }: Glp1CoachProps) {
   const { t } = useTranslation();
-  const { formatDate } = usePreferences();
   const medId = med.id;
   const glp1Drug = (
     med.custom_fields as { glp1_drug?: string } | null | undefined
@@ -63,201 +35,14 @@ export default function Glp1Coach({ med }: { med: Medication }) {
   // Oral/liquid GLP-1 (e.g. Rybelsus) still gets the PK curve, titration & fasting timer.
   const isInjectable = med.type_id === 'injection';
 
-  const sitesQ = useSiteSuggestion(medId);
   const curveQ = useSerumCurve(medId);
-  const pensQ = useMedicationPens(medId);
-  const injQ = useMedicationInjections(medId);
-  const titrationQ = useMedicationTitration(medId);
 
-  const suggestedSite = sitesQ.data?.suggestedSiteId ?? null;
-  const restingSites = useMemo(
-    () => new Set(sitesQ.data?.restingSiteIds ?? []),
-    [sitesQ.data]
-  );
-  // Honor the user's customized active site set (Settings → Customize sites).
-  const mapSites = useMemo(() => {
-    const active = sitesQ.data?.activeSiteIds;
-    if (!active || active.length === 0) return undefined;
-    const order = new Map(active.map((id, i) => [id, i] as const));
-    return INJECTION_SITES.filter((s) => order.has(s.id)).sort(
-      (a, b) => (order.get(a.id) ?? 0) - (order.get(b.id) ?? 0)
-    );
-  }, [sitesQ.data]);
-
-  const [selectedSite, setSelectedSite] = useState<string | null>(null);
-  const site = selectedSite ?? suggestedSite;
-  const inUsePen = pensQ.data?.find(
-    (p) => p.status === 'in_use' || p.status === 'sealed'
-  );
-
-  // Dose, date/time, and pen are user-editable at log time (mid-titration the dose changes,
-  // and shots are sometimes logged after the fact).
-  const [doseMg, setDoseMg] = useState(
-    med.dose_amount != null ? String(med.dose_amount) : ''
-  );
-  const [injectedAt, setInjectedAt] = useState('');
-  const [injectionNotes, setInjectionNotes] = useState('');
-  // null = "auto" (fall back to the in-use pen); 'none' = don't deduct; otherwise a pen id.
-  const [penChoice, setPenChoice] = useState<string | null>(null);
-  const effectivePenId = penChoice ?? inUsePen?.id ?? 'none';
-
-  const logMutation = useLogInjectionMutation(medId);
-  const addPenMutation = useCreatePenMutation(medId);
-  const deleteInjMutation = useDeleteInjectionMutation(medId);
-
-  // Add Inventory Form States
-  const [inventoryOpen, setInventoryOpen] = useState(false);
-  const [kind, setKind] = useState<'pen' | 'vial'>('pen');
-  const [label, setLabel] = useState('');
-  const [inventoryDoseMg, setInventoryDoseMg] = useState(
-    med.dose_amount != null ? String(med.dose_amount) : ''
-  );
-  const [concentration, setConcentration] = useState('');
-  const [volume, setVolume] = useState('');
-  const [dosesTotal, setDosesTotal] = useState('4');
-  const [openedAt, setOpenedAt] = useState('');
-  const [expiryDate, setExpiryDate] = useState('');
-  const [reorderFlag, setReorderFlag] = useState(false);
-  const [reorderThreshold, setReorderThreshold] = useState('1');
-  const [notes, setNotes] = useState('');
-
-  const preferencesContext = usePreferences();
-  const timezone =
-    preferencesContext?.timezone ||
-    Intl.DateTimeFormat().resolvedOptions().timeZone;
-  const today = todayInZone(timezone);
-
-  const calculatedBudDate = useMemo(() => {
-    if (!openedAt) return '';
-    try {
-      return addDays(openedAt, 28);
-    } catch {
-      return '';
-    }
-  }, [openedAt]);
-
-  const getExpiryStatus = (
-    targetDateStr: string | null,
-    currentDateStr: string
-  ) => {
-    if (!targetDateStr) return null;
-    try {
-      const target = new Date(targetDateStr + 'T00:00:00');
-      const current = new Date(currentDateStr + 'T00:00:00');
-      const diffDays =
-        (target.getTime() - current.getTime()) / (1000 * 60 * 60 * 24);
-      if (diffDays < 0) return 'expired';
-      if (diffDays <= 7) return 'near';
-      return 'good';
-    } catch {
-      return null;
-    }
-  };
-
-  const handleLog = () => {
-    const willDeduct = effectivePenId !== 'none';
-    logMutation.mutate(
-      {
-        medication_id: medId,
-        site,
-        dose_mg: doseMg ? Number(doseMg) : (med.dose_amount ?? null),
-        injected_at: injectedAt
-          ? new Date(injectedAt).toISOString()
-          : undefined,
-        pen_id: willDeduct ? effectivePenId : null,
-        deduct_pen: willDeduct,
-        notes: injectionNotes.trim() || null,
-      },
-      {
-        onSuccess: () => {
-          setSelectedSite(null);
-          setInjectedAt('');
-          setInjectionNotes('');
-        },
-      }
-    );
-  };
-
-  const handleCreateInventory = () => {
-    addPenMutation.mutate(
-      {
-        kind,
-        label: label.trim() || null,
-        dose_mg: inventoryDoseMg ? Number(inventoryDoseMg) : null,
-        concentration_mg_ml:
-          kind === 'vial' && concentration ? Number(concentration) : null,
-        volume_ml: kind === 'vial' && volume ? Number(volume) : null,
-        doses_total: dosesTotal ? Number(dosesTotal) : null,
-        opened_at: openedAt || null,
-        expiry_date: expiryDate || null,
-        bud_date: calculatedBudDate || null,
-        reorder_flag: reorderFlag,
-        reorder_threshold:
-          reorderFlag && reorderThreshold ? Number(reorderThreshold) : null,
-        notes: notes.trim() || null,
-        status: openedAt ? 'in_use' : 'sealed',
-      } as Partial<MedicationPen>,
-      {
-        onSuccess: () => {
-          setInventoryOpen(false);
-          // Reset fields
-          setKind('pen');
-          setLabel('');
-          setInventoryDoseMg(
-            med.dose_amount != null ? String(med.dose_amount) : ''
-          );
-          setConcentration('');
-          setVolume('');
-          setDosesTotal('4');
-          setOpenedAt('');
-          setExpiryDate('');
-          setReorderFlag(false);
-          setReorderThreshold('1');
-          setNotes('');
-        },
-      }
-    );
-  };
-
-  // Titration-step add form.
-  const addStepMutation = useAddTitrationStepMutation(medId);
-  const deleteStepMutation = useDeleteTitrationStepMutation(medId);
-  const [showAddStep, setShowAddStep] = useState(false);
-  const [stepDose, setStepDose] = useState('');
-  const [stepUnit, setStepUnit] = useState('mg');
-  const [stepStart, setStepStart] = useState('');
-  const [stepWeeks, setStepWeeks] = useState('');
-  const [stepStatus, setStepStatus] = useState<'planned' | 'active' | 'done'>(
-    'planned'
-  );
-  const [stepIsTaper, setStepIsTaper] = useState(false);
-
-  const handleAddStep = () => {
-    if (!stepDose) return;
-    addStepMutation.mutate(
-      {
-        dose_mg: Number(stepDose),
-        dose_unit: stepUnit || 'mg',
-        start_date: stepStart || null,
-        planned_weeks: stepWeeks ? Number(stepWeeks) : null,
-        status: stepStatus,
-        is_taper: stepIsTaper,
-      },
-      {
-        onSuccess: () => {
-          setShowAddStep(false);
-          setStepDose('');
-          setStepStart('');
-          setStepWeeks('');
-        },
-      }
-    );
-  };
-
-  const chartData = (curveQ.data?.curve ?? []).map((p) => ({
-    day: Number(p.day.toFixed(1)),
-    pct: Math.round(p.fraction * 100),
-  }));
+  const chartData = useMemo(() => {
+    return (curveQ.data?.curve ?? []).map((p) => ({
+      day: Number(p.day.toFixed(1)),
+      pct: Math.round(p.fraction * 100),
+    }));
+  }, [curveQ.data?.curve]);
 
   return (
     <div className="space-y-4">
@@ -265,170 +50,12 @@ export default function Glp1Coach({ med }: { med: Medication }) {
       {isOralGlp1 && <FastingTimer medId={medId} />}
 
       {/* Log injection + site rotation (injectable meds only) */}
-      {isInjectable && (
-        <Card>
-          <CardHeader>
-            <CardTitle className="flex items-center gap-2 text-base">
-              <Syringe className="h-4 w-4 text-blue-500" />{' '}
-              {t('medications.glp1.logInjection', 'Log injection')}
-            </CardTitle>
-          </CardHeader>
-          <CardContent className="space-y-4">
-            <div>
-              <div className="flex items-center justify-between gap-2">
-                <Label className="text-xs text-muted-foreground">
-                  {t('medications.glp1.injectionSite', 'Injection site')} ·{' '}
-                  <span className="text-green-600">
-                    {t('medications.glp1.legendSuggested', 'green = suggested')}
-                  </span>
-                  ,{' '}
-                  <span className="text-amber-600">
-                    {t(
-                      'medications.glp1.legendResting',
-                      'amber = resting <{{days}}d',
-                      { days: sitesQ.data?.restDays ?? 7 }
-                    )}
-                  </span>
-                </Label>
-                <InjectionSiteSettings />
-              </div>
-              <div className="mt-2 flex flex-col items-center gap-2">
-                <InjectionSiteBodyMap
-                  sites={mapSites}
-                  selectedSiteId={site}
-                  suggestedSiteId={suggestedSite}
-                  restingSiteIds={sitesQ.data?.restingSiteIds ?? []}
-                  onSelect={setSelectedSite}
-                />
-                <div className="flex items-center gap-2 text-sm">
-                  <span className="text-muted-foreground">
-                    {t('medications.glp1.selected', 'Selected:')}
-                  </span>
-                  <span className="font-medium">
-                    {site
-                      ? t(
-                          'medications.sites.label.' + site,
-                          INJECTION_SITES.find((s) => s.id === site)?.label ??
-                            site
-                        )
-                      : t('medications.glp1.tapZone', 'Tap a zone')}
-                  </span>
-                </div>
-              </div>
-              {site && restingSites.has(site) && (
-                <p className="mt-2 flex items-center gap-1 text-xs text-amber-600">
-                  <AlertTriangle className="h-3 w-3" />{' '}
-                  {t(
-                    'medications.glp1.lipoWarning',
-                    'This site was used recently — rotate to avoid lipohypertrophy.'
-                  )}
-                </p>
-              )}
-            </div>
-
-            <div className="grid grid-cols-2 gap-3">
-              <div className="space-y-1">
-                <Label className="text-xs text-muted-foreground">
-                  {t('medications.glp1.doseMg', 'Dose (mg)')}
-                </Label>
-                <Input
-                  type="number"
-                  step="0.05"
-                  value={doseMg}
-                  onChange={(e) => setDoseMg(e.target.value)}
-                  placeholder={
-                    med.dose_amount != null ? String(med.dose_amount) : '0'
-                  }
-                />
-              </div>
-              <div className="space-y-1">
-                <Label className="text-xs text-muted-foreground">
-                  {t('medications.glp1.dateTime', 'Date & time')}
-                </Label>
-                <Input
-                  type="datetime-local"
-                  value={injectedAt}
-                  onChange={(e) => setInjectedAt(e.target.value)}
-                />
-              </div>
-            </div>
-
-            <div className="space-y-1">
-              <Label className="text-xs text-muted-foreground">
-                {t('medications.glp1.deduct', 'Deduct from pen/vial')}
-              </Label>
-              <Select
-                value={effectivePenId}
-                onValueChange={(v) => setPenChoice(v)}
-              >
-                <SelectTrigger>
-                  <SelectValue
-                    placeholder={t(
-                      'medications.glp1.dontDeduct',
-                      "Don't deduct"
-                    )}
-                  />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="none">
-                    {t('medications.glp1.dontDeduct', "Don't deduct")}
-                  </SelectItem>
-                  {(pensQ.data ?? [])
-                    .filter((p) => p.status !== 'finished')
-                    .map((p) => (
-                      <SelectItem key={p.id} value={p.id}>
-                        {p.kind}
-                        {p.dose_mg ? ` ${p.dose_mg}mg` : ''} ·{' '}
-                        {(p.doses_total ?? 0) - p.doses_used} left
-                      </SelectItem>
-                    ))}
-                </SelectContent>
-              </Select>
-            </div>
-
-            <div className="space-y-1">
-              <Label className="text-xs text-muted-foreground">
-                {t('medications.glp1.notes', 'Notes')}
-              </Label>
-              <Input
-                value={injectionNotes}
-                onChange={(e) => setInjectionNotes(e.target.value)}
-                placeholder={t(
-                  'medications.glp1.notesPlaceholder',
-                  'e.g. Mild stinging, felt fine'
-                )}
-              />
-            </div>
-
-            <Button
-              onClick={handleLog}
-              disabled={!site || logMutation.isPending}
-              className="w-full"
-            >
-              {logMutation.isPending
-                ? t('medications.common.logging', 'Logging…')
-                : site
-                  ? t(
-                      'medications.glp1.logInjectionAt',
-                      'Log injection — {{site}}',
-                      {
-                        site: t(
-                          'medications.sites.label.' + site,
-                          INJECTION_SITES.find((s) => s.id === site)?.label ??
-                            site
-                        ),
-                      }
-                    )
-                  : t('medications.glp1.logInjection', 'Log injection')}
-            </Button>
-          </CardContent>
-        </Card>
-      )}
+      {isInjectable && <Glp1LogInjection med={med} />}
 
       {/* PK serum curve */}
       <Card>
         <CardHeader>
-          <CardTitle className="flex items-center justify-between text-base">
+          <CardTitle className="flex items-center justify-between text-base font-semibold">
             <span>
               {t('medications.glp1.pkTitle', 'PK serum level')} —{' '}
               {curveQ.data?.drugId ?? med.name}
@@ -502,570 +129,13 @@ export default function Glp1Coach({ med }: { med: Medication }) {
       </Card>
 
       {/* Pen / vial inventory (injectable meds only) */}
-      {isInjectable && (
-        <Card>
-          <CardHeader>
-            <CardTitle className="flex items-center justify-between text-base">
-              <span>
-                {t('medications.glp1.penInventory', 'Pen / vial inventory')}
-              </span>
-              <Button
-                variant="outline"
-                size="sm"
-                onClick={() => setInventoryOpen(true)}
-              >
-                {t('medications.glp1.addInventory', 'Add Inventory')}
-              </Button>
-            </CardTitle>
-          </CardHeader>
-          <CardContent className="space-y-2">
-            {(pensQ.data ?? []).length === 0 && (
-              <p className="text-sm text-muted-foreground">
-                {t('medications.glp1.noPens', 'No pens/vials tracked.')}
-              </p>
-            )}
-            {(pensQ.data ?? []).map((p) => {
-              const total = p.doses_total ?? 0;
-              const left = Math.max(0, total - p.doses_used);
-              const pct = total > 0 ? Math.round((left / total) * 100) : 0;
-              const low = total > 0 && pct <= 25;
-              const expStatus = getExpiryStatus(p.expiry_date, today);
-              const budStatus = getExpiryStatus(p.bud_date, today);
-
-              return (
-                <div key={p.id} className="rounded-lg border p-3 text-sm">
-                  <div className="flex items-center justify-between">
-                    <div className="flex items-center gap-2">
-                      <span className="font-medium capitalize">{p.kind}</span>
-                      {p.label && (
-                        <span className="text-muted-foreground">
-                          ({p.label})
-                        </span>
-                      )}
-                      {p.dose_mg ? (
-                        <span className="text-muted-foreground">
-                          {p.dose_mg} mg
-                        </span>
-                      ) : null}
-                      {p.concentration_mg_ml ? (
-                        <span className="text-muted-foreground">
-                          · {p.concentration_mg_ml} mg/mL
-                        </span>
-                      ) : null}
-                      {p.status === 'in_use' && (
-                        <Badge variant="secondary" className="text-[10px]">
-                          in use
-                        </Badge>
-                      )}
-                    </div>
-                    <div className="flex items-center gap-2">
-                      {p.reorder_flag &&
-                        p.reorder_threshold != null &&
-                        left <= p.reorder_threshold && (
-                          <Badge
-                            variant="destructive"
-                            className="flex items-center gap-1 text-[10px]"
-                          >
-                            <AlertTriangle className="h-3 w-3" /> Reorder
-                          </Badge>
-                        )}
-                      <span className="font-medium tabular-nums">
-                        {left}/{total || '?'}{' '}
-                        <span className="font-normal text-muted-foreground">
-                          doses
-                        </span>
-                      </span>
-                    </div>
-                  </div>
-                  <div className="mt-2 h-2 w-full overflow-hidden rounded-full bg-muted">
-                    <div
-                      className={`h-full rounded-full transition-all ${low ? 'bg-amber-500' : 'bg-blue-500'}`}
-                      style={{ width: `${pct}%` }}
-                    />
-                  </div>
-                  {(p.expiry_date || p.bud_date) && (
-                    <div className="mt-1.5 flex gap-3 text-xs text-muted-foreground flex-wrap items-center">
-                      {p.expiry_date && (
-                        <span className="flex items-center gap-1">
-                          Exp {p.expiry_date}
-                          {expStatus === 'expired' && (
-                            <Badge
-                              variant="destructive"
-                              className="text-[9px] px-1 py-0 h-4"
-                            >
-                              Expired
-                            </Badge>
-                          )}
-                          {expStatus === 'near' && (
-                            <Badge className="text-[9px] px-1 py-0 h-4 bg-amber-500 text-white hover:bg-amber-600">
-                              Near Exp
-                            </Badge>
-                          )}
-                        </span>
-                      )}
-                      {p.bud_date && (
-                        <span className="flex items-center gap-1">
-                          BUD {p.bud_date}
-                          {budStatus === 'expired' && (
-                            <Badge
-                              variant="destructive"
-                              className="text-[9px] px-1 py-0 h-4"
-                            >
-                              Expired (BUD)
-                            </Badge>
-                          )}
-                          {budStatus === 'near' && (
-                            <Badge className="text-[9px] px-1 py-0 h-4 bg-amber-500 text-white hover:bg-amber-600 font-semibold">
-                              BUD Warning
-                            </Badge>
-                          )}
-                        </span>
-                      )}
-                    </div>
-                  )}
-                  {p.notes && (
-                    <p className="mt-1.5 text-xs text-muted-foreground italic border-t pt-1">
-                      Notes: {p.notes}
-                    </p>
-                  )}
-                </div>
-              );
-            })}
-          </CardContent>
-        </Card>
-      )}
+      {isInjectable && <Glp1InventoryManager med={med} />}
 
       {/* Titration plan */}
-      <Card>
-        <CardHeader>
-          <div className="flex items-center justify-between">
-            <CardTitle className="text-base">
-              {t('medications.glp1.titrationTitle', 'Dose titration plan')}
-            </CardTitle>
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={() => setShowAddStep((s) => !s)}
-            >
-              <Plus className="mr-1 h-3.5 w-3.5" />{' '}
-              {t('medications.glp1.addStep', 'Add step')}
-            </Button>
-          </div>
-          <p className="text-xs text-muted-foreground">
-            Your planned GLP-1 dose escalation (e.g. 0.25 → 0.5 → 1.0 mg). Add
-            each step with its start date; mark one “active” as your current
-            dose.
-          </p>
-        </CardHeader>
-        <CardContent>
-          {showAddStep && (
-            <div className="mb-4 space-y-2 rounded-md border p-3">
-              <div className="grid grid-cols-2 gap-2">
-                <div className="space-y-1">
-                  <Label className="text-xs text-muted-foreground">Dose</Label>
-                  <div className="flex gap-1">
-                    <Input
-                      type="number"
-                      step="0.05"
-                      value={stepDose}
-                      onChange={(e) => setStepDose(e.target.value)}
-                      placeholder="1.0"
-                    />
-                    <Input
-                      className="w-16"
-                      value={stepUnit}
-                      onChange={(e) => setStepUnit(e.target.value)}
-                    />
-                  </div>
-                </div>
-                <div className="space-y-1">
-                  <Label className="text-xs text-muted-foreground">
-                    Start date
-                  </Label>
-                  <Input
-                    type="date"
-                    value={stepStart}
-                    onChange={(e) => setStepStart(e.target.value)}
-                  />
-                </div>
-                <div className="space-y-1">
-                  <Label className="text-xs text-muted-foreground">
-                    Weeks at this dose
-                  </Label>
-                  <Input
-                    type="number"
-                    value={stepWeeks}
-                    onChange={(e) => setStepWeeks(e.target.value)}
-                    placeholder="4"
-                  />
-                </div>
-                <div className="space-y-1">
-                  <Label className="text-xs text-muted-foreground">
-                    Status
-                  </Label>
-                  <Select
-                    value={stepStatus}
-                    onValueChange={(v) =>
-                      setStepStatus(v as 'planned' | 'active' | 'done')
-                    }
-                  >
-                    <SelectTrigger>
-                      <SelectValue />
-                    </SelectTrigger>
-                    <SelectContent>
-                      <SelectItem value="planned">
-                        {t('medications.glp1.statusPlanned', 'Planned')}
-                      </SelectItem>
-                      <SelectItem value="active">
-                        {t('medications.glp1.statusActive', 'Active (current)')}
-                      </SelectItem>
-                      <SelectItem value="done">
-                        {t('medications.glp1.statusDone', 'Done')}
-                      </SelectItem>
-                    </SelectContent>
-                  </Select>
-                </div>
-              </div>
-              <div className="flex items-center justify-between">
-                <label className="flex items-center gap-2 text-xs text-muted-foreground">
-                  <Switch
-                    checked={stepIsTaper}
-                    onCheckedChange={setStepIsTaper}
-                  />{' '}
-                  Taper (dose reduction)
-                </label>
-                <div className="flex gap-2">
-                  <Button
-                    variant="ghost"
-                    size="sm"
-                    onClick={() => setShowAddStep(false)}
-                  >
-                    Cancel
-                  </Button>
-                  <Button
-                    size="sm"
-                    onClick={handleAddStep}
-                    disabled={!stepDose || addStepMutation.isPending}
-                  >
-                    Add
-                  </Button>
-                </div>
-              </div>
-            </div>
-          )}
-          {(titrationQ.data ?? []).length === 0 ? (
-            <p className="text-sm text-muted-foreground">
-              {t(
-                'medications.glp1.noTitration',
-                'No titration steps yet — tap “Add step” to plan your dose escalation.'
-              )}
-            </p>
-          ) : (
-            <ol className="relative space-y-4 border-l border-muted pl-6">
-              {(titrationQ.data ?? []).map((step) => {
-                const active = step.status === 'active';
-                const done = step.status === 'done';
-                return (
-                  <li key={step.id} className="relative">
-                    <span
-                      className={`absolute -left-[27px] mt-1 h-3.5 w-3.5 rounded-full border-2 ${
-                        active
-                          ? 'border-blue-500 bg-blue-500'
-                          : done
-                            ? 'border-green-500 bg-green-500'
-                            : 'border-muted-foreground/40 bg-background'
-                      }`}
-                    />
-                    <div className="flex items-center justify-between text-sm">
-                      <div>
-                        <span className="font-medium">
-                          {step.dose_mg} {step.dose_unit}
-                        </span>
-                        {step.start_date && (
-                          <span className="text-muted-foreground">
-                            {' '}
-                            · {step.start_date}
-                          </span>
-                        )}
-                        {step.planned_weeks ? (
-                          <span className="text-muted-foreground">
-                            {' '}
-                            · {step.planned_weeks} wks
-                          </span>
-                        ) : null}
-                        {step.is_taper && (
-                          <Badge variant="outline" className="ml-2 text-[10px]">
-                            taper
-                          </Badge>
-                        )}
-                      </div>
-                      <div className="flex items-center gap-1">
-                        <Badge
-                          variant={active ? 'default' : 'secondary'}
-                          className="capitalize"
-                        >
-                          {step.status}
-                        </Badge>
-                        <Button
-                          variant="ghost"
-                          size="icon"
-                          className="h-6 w-6 text-muted-foreground hover:text-destructive"
-                          onClick={() => deleteStepMutation.mutate(step.id)}
-                          aria-label="Delete titration step"
-                        >
-                          <Trash2 className="h-3 w-3" />
-                        </Button>
-                      </div>
-                    </div>
-                  </li>
-                );
-              })}
-            </ol>
-          )}
-        </CardContent>
-      </Card>
+      <Glp1TitrationManager med={med} />
 
       {/* Recent injections (injectable meds only) */}
-      {isInjectable && (
-        <Card>
-          <CardHeader>
-            <CardTitle className="text-base">
-              {t('medications.glp1.recentInjections', 'Recent injections')}
-            </CardTitle>
-          </CardHeader>
-          <CardContent className="space-y-2">
-            {(injQ.data ?? []).length === 0 && (
-              <p className="text-sm text-muted-foreground">
-                {t(
-                  'medications.glp1.noInjections',
-                  'No injections logged yet.'
-                )}
-              </p>
-            )}
-            {(injQ.data ?? []).slice(0, 8).map((inj) => (
-              <div
-                key={inj.id}
-                className="flex items-center justify-between rounded-md border p-2 text-sm gap-2"
-              >
-                <div className="flex flex-col min-w-0">
-                  <span className="font-medium truncate">
-                    {INJECTION_SITES.find((s) => s.id === inj.site)?.label ??
-                      inj.site ??
-                      '—'}
-                  </span>
-                  <span className="text-xs text-muted-foreground">
-                    {inj.dose_mg ? `${inj.dose_mg} mg · ` : ''}
-                    {formatDate(inj.injected_at)}
-                  </span>
-                  {inj.notes && (
-                    <span className="text-xs text-muted-foreground italic mt-0.5 block truncate">
-                      Note: {inj.notes}
-                    </span>
-                  )}
-                </div>
-                <Button
-                  variant="ghost"
-                  size="icon"
-                  className="h-7 w-7 text-muted-foreground hover:text-destructive shrink-0"
-                  onClick={() => deleteInjMutation.mutate(inj.id)}
-                  disabled={deleteInjMutation.isPending}
-                  aria-label="Delete injection entry"
-                >
-                  <Trash2 className="h-3.5 w-3.5" />
-                </Button>
-              </div>
-            ))}
-          </CardContent>
-        </Card>
-      )}
-      {/* Add Inventory Dialog */}
-      <Dialog open={inventoryOpen} onOpenChange={setInventoryOpen}>
-        <DialogContent className="max-w-md">
-          <DialogHeader>
-            <DialogTitle>Add Pen / Vial Inventory</DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4 py-2">
-            <div className="grid grid-cols-2 gap-3">
-              <div className="space-y-1">
-                <Label htmlFor="inv-kind">Kind</Label>
-                <Select
-                  value={kind}
-                  onValueChange={(v) => {
-                    setKind(v as 'pen' | 'vial');
-                    if (v === 'pen') {
-                      setDosesTotal('4');
-                    } else {
-                      setDosesTotal('10');
-                    }
-                  }}
-                >
-                  <SelectTrigger id="inv-kind">
-                    <SelectValue />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="pen">Pen</SelectItem>
-                    <SelectItem value="vial">Vial</SelectItem>
-                  </SelectContent>
-                </Select>
-              </div>
-
-              <div className="space-y-1">
-                <Label htmlFor="inv-label">Label / Name</Label>
-                <Input
-                  id="inv-label"
-                  placeholder="e.g. Pen #2, Vial Batch A"
-                  value={label}
-                  onChange={(e) => setLabel(e.target.value)}
-                />
-              </div>
-            </div>
-
-            <div className="grid grid-cols-2 gap-3">
-              <div className="space-y-1">
-                <Label htmlFor="inv-dose">Dose Strength (mg)</Label>
-                <Input
-                  id="inv-dose"
-                  type="number"
-                  step="0.05"
-                  value={inventoryDoseMg}
-                  onChange={(e) => setInventoryDoseMg(e.target.value)}
-                  placeholder={
-                    med.dose_amount != null ? String(med.dose_amount) : '0'
-                  }
-                />
-              </div>
-
-              <div className="space-y-1">
-                <Label htmlFor="inv-doses-total">Total Doses</Label>
-                <Input
-                  id="inv-doses-total"
-                  type="number"
-                  value={dosesTotal}
-                  onChange={(e) => setDosesTotal(e.target.value)}
-                />
-              </div>
-            </div>
-
-            {kind === 'vial' && (
-              <div className="grid grid-cols-2 gap-3">
-                <div className="space-y-1">
-                  <Label htmlFor="inv-concentration">
-                    Concentration (mg/mL)
-                  </Label>
-                  <Input
-                    id="inv-concentration"
-                    type="number"
-                    step="0.1"
-                    placeholder="e.g. 5"
-                    value={concentration}
-                    onChange={(e) => setConcentration(e.target.value)}
-                  />
-                </div>
-
-                <div className="space-y-1">
-                  <Label htmlFor="inv-volume">Volume (mL)</Label>
-                  <Input
-                    id="inv-volume"
-                    type="number"
-                    step="0.1"
-                    placeholder="e.g. 2"
-                    value={volume}
-                    onChange={(e) => setVolume(e.target.value)}
-                  />
-                </div>
-              </div>
-            )}
-
-            <div className="grid grid-cols-2 gap-3">
-              <div className="space-y-1">
-                <Label htmlFor="inv-opened">Date Opened</Label>
-                <Input
-                  id="inv-opened"
-                  type="date"
-                  value={openedAt}
-                  onChange={(e) => setOpenedAt(e.target.value)}
-                />
-              </div>
-
-              <div className="space-y-1">
-                <Label htmlFor="inv-expiry">Expiry Date</Label>
-                <Input
-                  id="inv-expiry"
-                  type="date"
-                  value={expiryDate}
-                  onChange={(e) => setExpiryDate(e.target.value)}
-                />
-              </div>
-            </div>
-
-            {openedAt && (
-              <div className="rounded-md bg-muted p-2 text-xs">
-                <span className="font-semibold text-muted-foreground">
-                  Calculated Beyond-Use Date (BUD):
-                </span>{' '}
-                <span className="font-medium">
-                  {calculatedBudDate || 'N/A'}
-                </span>
-                <p className="text-[10px] text-muted-foreground mt-0.5">
-                  Based on 28-day stability window from first opening.
-                </p>
-              </div>
-            )}
-
-            <div className="rounded-md border p-3 space-y-2">
-              <div className="flex items-center justify-between">
-                <Label htmlFor="inv-reorder" className="flex flex-col gap-0.5">
-                  <span>Enable Reorder Warning</span>
-                  <span className="font-normal text-[10px] text-muted-foreground">
-                    Alert when remaining doses are low
-                  </span>
-                </Label>
-                <Switch
-                  id="inv-reorder"
-                  checked={reorderFlag}
-                  onCheckedChange={setReorderFlag}
-                />
-              </div>
-
-              {reorderFlag && (
-                <div className="space-y-1 pt-1">
-                  <Label htmlFor="inv-threshold">
-                    Reorder Threshold (doses left)
-                  </Label>
-                  <Input
-                    id="inv-threshold"
-                    type="number"
-                    value={reorderThreshold}
-                    onChange={(e) => setReorderThreshold(e.target.value)}
-                  />
-                </div>
-              )}
-            </div>
-
-            <div className="space-y-1">
-              <Label htmlFor="inv-notes">Notes</Label>
-              <Textarea
-                id="inv-notes"
-                placeholder="Batch number, brand, pharmacy info..."
-                rows={2}
-                value={notes}
-                onChange={(e) => setNotes(e.target.value)}
-              />
-            </div>
-          </div>
-          <DialogFooter>
-            <Button variant="ghost" onClick={() => setInventoryOpen(false)}>
-              Cancel
-            </Button>
-            <Button
-              onClick={handleCreateInventory}
-              disabled={addPenMutation.isPending}
-            >
-              {addPenMutation.isPending ? 'Saving...' : 'Add Inventory'}
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+      {isInjectable && <Glp1RecentInjections med={med} />}
     </div>
   );
 }
